requirements.txt
google-genai>=0.1.0
python-dotenv>=1.0.1
customtkinter>=5.2.2
